commit tugas6

+++ src/main/java/com/tugas5/apotek/services/CategoryService.java
@@ -20,8 +20,29 @@ public class CategoryService {
         return categoryRepository.save(category);
     }
 
-    public void deleteCategory(Integer id) {
+    public boolean canDeleteCategory(Integer id) {
+        Category category = getCategoryById(id);
+        return category.getObats().isEmpty();
+    }
+
+    public boolean deleteCategory(Integer id) {
+        if (!canDeleteCategory(id)) {
+            return false;
+        }
         categoryRepository.deleteById(id);
+        return true;
+    }
+
+    public Category getCategoryById(Integer id) {
+        return categoryRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Category not found"));
+    }
+
+    public Category updateCategory(Integer id, Category category) {
+        Category existingCategory = getCategoryById(id);
+        existingCategory.setName(category.getName());
+        existingCategory.setDescription(category.getDescription());
+        return categoryRepository.save(existingCategory);
     }
 }
